refactor campaign definition detail router
